food_distribution_system/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.0.1
Flask-Cors==3.0.10
Werkzeug==2.2.3
python-dotenv==0.20.0
opencv-python==4.7.0.72  # For image handling
